Fixed `_calc_value` edge case for fixed metabolites
In some cases control coefficients will have an expression of `1`
when the numerator and denominator are the same (this happens in
some cases where there are fixed metabolites. To prevent this
if a value length of 0 for the number of symbols (ecs, fluxes, and
species) is detected then the `subsdict` in `_calc_value` and
`_calc_abs_value` is set to `None` which forces control patterns
to create their own `subsdict` dictionaries. This fixes the
problem but leads to values of less than one for the control
coefficient value (even though the expression is 1).