Reorganise `do_symca` code in a more logical form
Previously it was much more difficult to understand what was going
on in this method. There were two logical statements testing for
`internal_fixed` and now there is only a single one. Some code was
refactored to static methods (which will later be refactored to
functions....) and moved to `symca_toolbox`. New methods are
`make_internals_dic` and `make_CC_dot_dict`. These changes
necessitate a refactoring/rewriting of the save/load methods which
will be included in the next commit (hopefully).