Add the `extract_model` function
The `extract_model` function tries to get the `mod` field of any
object and checks if it is an instance of `PysMod`. It then
returns this `PysMod` object together with a string representation
of the type of the original object. In the case where the object
itself is an instance of `PysMod`, the object itself will be
returned together with a string `"PysMod"`. An assertion error
will occur if the object to be returned is not `PysMod`.