Add cc expresions fixing for `internal_fixed` = True
Expressions are "fixed" by dividing both numerators and the
common denominator by species involved in a moiety conserved cycle.
However when `internal_fixed` is True additional factorisation
takes place (to find different common denominators, one for each
block and additionally one for the `1` common denominator). This
undoes the original fixing.

This commit adds back the fixing but also makes sure that it only
applies it for reactions where it actually makes sense. e.g. when
a block is isolated from the conserved moieties (with members S1
and S2), its control coefficients will not have their numerators
and denominators divided by S1*S2 but the block containing these
species will.

Finally: In the case where the numerator expression is `0` or `1`
(which corresponds with a `1` value for the denominator), there
will be a single control pattern corresponding to this value.
This does not really matter in reality but it does ensure that
things are consistant.