Revert previous change (addition of print statement)
Previous commit (855cccc61f82461da10e977934c4bdcfc870df5a) had an
incomplete log, but was supposed to indicate when additional
simplification was taking place in the case of fixed internal
metabolites with a print statement. However, the printed message
only appeared after the method call was complete and was therefore
useless. The problem went unnoticed for small models but was very
obvious as soon as bigger models were tested. This will be
investigated and fixed in the future.