Fix symca expressions for branched systems
Previously, control coefficient expressions for branched systems
were not rearranged to the correct form - unless the system also
contained a moiety conserved cycle. The algorythm was previously
changed to only "fix" expressions when they needed fixing which
applied specifically to systems where there were internal fixed
metabolites. This was done to prevent parts of the system that
were isolated from a conserved moiety to include these species in
their expressions (in the control pattern numerator). However
as a side effect branched system expressions were not fixed.
Expression fixing consists of two steps: (1) Numerators are
multiplied by the common denominator's denominator and (2) the
numerators are divided by the "fixed denom" (which consists of
symbols for certain species in the system). Previously both steps
were applied on a case by case basis, now step 1 is applied to all
numerators before testing for the conditions for, and the
application of, step 2.