Streamline evaluation of symbolic expressions
In an early version of `Symca` the control coefficients were
evaluated using the built in `sympy` method `subs`. After
ThermoKin was developed this was changed to an `eval` based method
which was an order of magnitude (or more) faster than the `sympy`
method. Unfortunately this method has some limitations - it cannot
evaluate any mathematical operators other than +,-,/, * and **.
Therefore any equations with something like a log will cause an
exception. Now a generic `get_value` function is included that
first tries the `eval` based method and defualts to the `sympy`
method in the case of an exception. Code duplication between
`Symca` and `Thermokin` was also removed as these functions were
moved to misc.