Add automatic generation of gamma/keq terms
Gamma/Keq terms will now be generated and added to all ThermoKin
objects whenever they are not included in the definition file. In
the case where no definition file exists (or the `overwrite` flag
argument is `True`) these terms will be written to the reqn file
automatically as additional terms (specified with !G. Otherwise a
user needs to add them manually with a new `term_to_file` function
which will be made more accessable in a near future commit.