Add memory optimisation for plotting
Instead of keeping a large dictionary in memory in order to figure
out line categories, the dictionary is now created on the fly and
not saved. In this way a trade off between CPU and memory is made,
where the same object will need to be created whenever a `ScanFig`
object is instantiated by `Data2D.plot`. The justification here is
that a user will probably only create a single `ScanFig` from each
`Data2D` object. Therefore it's much more wasteful to let the
dictionary persist. The side effect of this optimisation is that
now a ScanFig object would keep the dictionary in its memory (instead
of referencing it). This issue was largely circumvented by chopping
down the dictionary inside `ScanFig` to only include the needed
elements.