Added automatic axvline and ability to specify status of categories
If the new `axvline` argument is True (default) a vertical line
will be drawn at current value of the attribute of self.mod that
corresponds to self.scan_results.scan_in (which in most cases except
for results generated with `Scan1` should be the init value of this
parameter). The `category_manifest` dictionary allows for
specification of the defualt status of any categories (True/False).
This does not necessarily affect the visibility of the lines
represented by a category, becuase these lines need all their
categories to be True to be visible.

Many other small formatting changes were made.