Add new category filters and defualt category statuses to par_scans
`ScanFigs` generated by `Data2D` objects resulting from
`do_par_scan` will now have certain categories switched on by
defualt. Elasticity scans also have new category filters -  Term
elasticities will not be visible unless their "parent" elasticity
is switched on (as well as the top level "Term Elasticities"
category). Another noteworthy changes include the addition of a
`_rname` attribute to `Term`. In the case of term elasticities,
the value will be the name of the "full elasticity" which it forms
part of and in the case of the "full elasticity" it will be the name
of the reaction it belongs to. Other small formatting changes were made.