Optimisation of parameter scan methods
When calculating the control coefficient value, the values of the
individual control patterns are summed. This is a legacy from before
eval was used to evaluate expressions (sympy subs was used). When
performing parameter scans where all the control pattern values and
the control coeffient value are needed this strategy means double
work. Worse still is when the percentage contribution of all the
control patterns are calculated as each calculation involves
calculating the value of every other control pattern as well which
means exponentially more work.

Here these shortcomings were addressed in the case of the
`do_par_scan` method of `CCoef`. Now the minimum number of
operations are performed leading to signigicant speedups for both
"percentage" and "value" scan modes.