Optimise load speed of symca using pickle based approach
Originally symca used a saving approach that was purely pickle based
that meant that data saved with previous versions of symca were
possibly incompatible with newer versions. A new json based approach
was used in favor of the original approach to circumvent the the
associated issues. Unfortunately in the json based approach,
sympy expressions were saved as strings and this turned out to result
in very slow loading of symca data for large models. A new pickle
based approach was used here in favor of json, but the optimisations
in terms of backward compatibility and portability was retained.
A 2-3 fold speed up was observed in the limited test cases. Unfortunately
this means that old data will have to be regenerated.