Fix strange Thermokin bug
In some cases it seems that dividing the rate equation by the mass
action term (to get the combined rate capacity and binding terms)
and subsequently factorising the result does not yield an expression
that is free from the Keq term (which it should). Adding another
identical factorisation step fixes this.