Make get_value and related functions more robust.
The get_value_eval and get_value_sympy functions are supposed to
be identical in the sense that they can take the same arguments
and return *very* similar outputs. Unfortunately, this was not
true and get_value_sympy could not handle subs_dicts with arrays
as values. Previously this was not an issue as get_value_sympy is
used only as a fallback and had thus this lack of functionality was
never encountered. Secondly, each of the two functions
indiscriminitely performs a type conversion that can potentially
be computationally expensive relative to conditionally applying the
conversion subsequent to a type check. A fix was therefore
implemented.