Fix d3networkx node adding functionality.
The older version of networkx handled adding of node attributes
in a slightly different manner compared to the new 2 series. Since
EventfulGraph is a child class of networkx.Graph, this change
caused node attributes to not be added when creating a new node.
When this empty node dictionary was passed to the javascript
component of the widget, it resulted in malformed d3 nodes being
created.

This fix simply copies the version 1.x series version of the
networkx.Graph.add_node method to EventfulGraph together with some
minor changes. It seems to work, but might need further testing.