Fix saving of ModelGraph svg images.
Somewhere a change occured that caused the svg code defining the
image to be wrapped in some additional xml cruft. This fix simply
cuts out the relavent svg from the original xml string and saves it
instead of the whole xml document. Ideally this change should have
occured in ForceDirectedGraph, however the svg_image attribute is empty
until the save() method of ModelGraph is called. To fix this without
digging in to the details too much, the fix was applied at the
ModelGraph level.